requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
x402>=0.2.1
openai>=1.30.0
python-dotenv>=1.0.0